pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "streamexpert-support-crm"
version = "0.1.0"
description = "Compliant Telegram IPTV support and CRM platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0",
  "apscheduler>=3.10.4",
  "fastapi>=0.115.0",
  "openai>=1.54.0",
  "python-dotenv>=1.0.1",
  "pydantic-settings>=2.6.0",
  "psycopg[binary]>=3.2.0",
  "asyncpg>=0.29.0",
  "redis>=5.1.0",
  "sqlalchemy>=2.0.36",
  "aiosqlite>=0.17.0",
  "uvicorn[standard]>=0.32.0",
  "telethon>=1.36.0",
  "praw>=7.7.1",
  "groq>=0.11.0",
  "google-generativeai>=0.8.3",
  "PySocks>=1.7.1",
  "emoji>=2.14.0",
  "sentry-sdk[fastapi]>=2.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.2",
  "pytest>=8.3.3",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
